Move Features from services to utils
