+++ apps/store/src/components/HeadSeoInfo/HeadSeoInfo.tsx
@@ -1,7 +1,7 @@
 import { ISbAlternateObject, ISbStoryData } from '@storyblok/react'
 import Head from 'next/head'
-import { Features } from '@/services/Features'
 import { SEOData } from '@/services/storyblok/storyblok'
+import { Features } from '@/utils/Features'
 import { isRoutingLocale, toIsoLocale } from '@/utils/l10n/localeUtils'
 import { ORIGIN_URL } from '@/utils/PageLink'
 import { StructuredDataOrganization } from './StructuredDataOrganization'
@@ -51,7 +51,7 @@ const AlternateLinks = ({ story }: { story: ISbStoryData<SEOData> }) => {
 }
 
 const isVisibleAlternate = (alternate: ISbAlternateObject) =>
-  Features.isEnabled('ENGLISH_LANGUAGE') || !getHrefLang(alternate.full_slug).startsWith('en-')
+  Features.enabled('ENGLISH_LANGUAGE') || !getHrefLang(alternate.full_slug).startsWith('en-')
 
 const AlternateLink = ({ fullSlug }: { fullSlug: string }) => {
   return (
