+++ apps/store/src/services/storyblok/storyblok.ts
@@ -58,9 +58,9 @@ import { VideoBlock } from '@/blocks/VideoBlock'
 import { VideoListBlock } from '@/blocks/VideoListBlock'
 import { fetchStory, StoryblokFetchParams } from '@/services/storyblok/Storyblok.helpers'
 import { isBrowser } from '@/utils/env'
+import { Features } from '@/utils/Features'
 import { getLocaleOrFallback, isRoutingLocale } from '@/utils/l10n/localeUtils'
 import { Language, RoutingLocale } from '@/utils/l10n/types'
-import { Features } from '../Features'
 import { GLOBAL_STORY_PROP_NAME, STORY_PROP_NAME } from './Storyblok.constant'
 
 export type SbBaseBlockProps<T> = {
@@ -311,7 +311,7 @@ export const getPageLinks = async (): Promise<PageLink[]> => {
     if (!isRoutingLocale(locale)) return
 
     const { language } = getLocaleOrFallback(locale)
-    if (language === Language.En && !Features.isEnabled('ENGLISH_LANGUAGE')) return
+    if (language === Language.En && !Features.enabled('ENGLISH_LANGUAGE')) return
 
     if (slugParts[0] === 'global') return
     pageLinks.push({
